refactor(service): move mapper package to service.service package #17
