+++ src/test/java/pl/gawor/tayckner/taycknerbackend/service/service/mapper/UserMapperTest.java
@@ -1,4 +1,4 @@
-package pl.gawor.tayckner.taycknerbackend.service.mapper;
+package pl.gawor.tayckner.taycknerbackend.service.service.mapper;
 
 import org.junit.jupiter.api.Test;
 import pl.gawor.tayckner.taycknerbackend.core.model.UserModel;
